requirements.txt
pyatuogui
mouse
cv2
pytesseract
pynpyt
pillow
